pyproject.toml
[tool.poetry]
name = "ec2-win-pkcs11"
version = "0.1.0"
description = ""
authors = ["Daniel Jones <tortxof@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
boto3 = "^1.12.25"
python-pkcs11 = "^0.6.0"
Click = "^7.1.1"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
